fix: delete the form-scope scan, whose delivers verdict is now wrong
The second review round found the scan is not merely callerless. Its
`delivers` field encodes the premise this branch removes:
`unboundFormDelivers` answers false for a form with no method or a
method="get" one, but a bound submitter now carries its own
formmethod="post", which overrides the form, so those forms DO deliver.
Its tests asserted the wrong values, and the blog fixture it would score
false is the one the e2e proves runs its action.

Salvaged the two live surfaces into form-action-holes.test.js:
classifyActionHole, which the form-action-not-a-get-action rule still
uses, and the enctype pin that keeps the renderer allowlist and the client
denylist from drifting.

Also cleared four more residuals of the same premise: the onError message
claimed a bound submitter's formmethod="get" is honoured when it is
refused at render, the renderer comment said a submission targets the page
rather than whatever the form targets, the streaming machine kept an
orphaned form-scope comment, and three doc surfaces called a refused
method="get" on a bound form a warning.
