--- packages/server/test/scanner/html-form-scopes.test.js
@@ -1,325 +1,0 @@
-import { test } from 'node:test';
-import assert from 'node:assert/strict';
-
-import {
-  scanHtmlFormScopes,
-  classifyActionHole,
-  matchClosingBrace,
-  extractWebComponentClassBodies,
-  redactStringsAndTemplates,
-} from '../../src/js-scan.js';
-import { PARSEABLE_ENCTYPES } from '../../../core/src/form-action.js';
-import { readFile } from 'node:fs/promises';
-
-/**
- * Unit tests for the lexical form-scope scan in `js-scan.js`, plus the shared
- * template primitives around it (`classifyActionHole`, `matchClosingBrace`,
- * `extractWebComponentClassBodies`) and the enctype constant the renderer and
- * the client guard both answer to. The shapes the scan must NOT read as markup
- * (a plain string, a `css` template, an HTML comment) are what keep any
- * consumer of it off a docs page.
- */
-
-test('a submitter reports the scope of its enclosing form', () => {
-  assert.deepEqual(
-    scanHtmlFormScopes('html`<form action=${save}><button formaction=${del}>x</button></form>`').submitters,
-    [{ tag: 'button', scope: 'bound', delivers: null, expr: 'del' }],
-  );
-  assert.deepEqual(
-    scanHtmlFormScopes('html`<form><button formaction=${del}>x</button></form>`').submitters,
-    [{ tag: 'button', scope: 'unbound', delivers: false, expr: 'del' }],
-  );
-  assert.deepEqual(
-    scanHtmlFormScopes('html`<button formaction=${del}>x</button>`').submitters,
-    [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }],
-  );
-});
-
-test('a nested template inherits the enclosing form scope', () => {
-  // What the renderer does: `render` threads `formScope` through arrays,
-  // `repeat`, and nested templates, so a per-row button in a bound form is
-  // bound.
-  const bound = 'html`<form action=${save}>${rows.map((r) => html`<button formaction=${del}>x</button>`)}</form>`';
-  assert.deepEqual(scanHtmlFormScopes(bound).submitters, [{ tag: 'button', scope: 'bound', delivers: null, expr: 'del' }]);
-  const unbound = 'html`<form>${rows.map((r) => html`<button formaction=${del}>x</button>`)}</form>`';
-  assert.deepEqual(scanHtmlFormScopes(unbound).submitters, [{ tag: 'button', scope: 'unbound', delivers: false, expr: 'del' }]);
-});
-
-test('the scope closes at </form> and does not leak forward', () => {
-  const src = 'html`<form action=${save}></form><button formaction=${del}>x</button>`';
-  assert.deepEqual(scanHtmlFormScopes(src).submitters, [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }]);
-});
-
-test('a start tag split across holes is still one tag', () => {
-  const src = 'html`<form action=${save} class="a"><button class=${c} formaction=${del}>x</button></form>`';
-  assert.deepEqual(scanHtmlFormScopes(src).submitters, [{ tag: 'button', scope: 'bound', delivers: null, expr: 'del' }]);
-});
-
-test('a `>` inside a quoted attribute value does not close the tag', () => {
-  const src = 'html`<form action=${save}><div title="a>b"></div><button formaction=${del}>x</button></form>`';
-  assert.deepEqual(scanHtmlFormScopes(src).submitters, [{ tag: 'button', scope: 'bound', delivers: null, expr: 'del' }]);
-});
-
-test('a custom-element start tag is reported, its close tag is not', () => {
-  assert.deepEqual(
-    scanHtmlFormScopes('html`<form action=${save}><row-btn></row-btn></form>`').tagUses,
-    [{ tag: 'row-btn', scope: 'bound', delivers: null, expr: null }],
-  );
-  assert.deepEqual(
-    scanHtmlFormScopes('html`<form><row-btn></row-btn></form>`').tagUses,
-    [{ tag: 'row-btn', scope: 'unbound', delivers: false, expr: null }],
-  );
-});
-
-test('only an html-tagged literal is read as markup', () => {
-  // The carve-out the whole rule rests on: the framework's own website renders
-  // `<form action=${fn}>` as a code SAMPLE.
-  assert.deepEqual(scanHtmlFormScopes("const s = '<form><button formaction=x></button></form>';").submitters, []);
-  assert.deepEqual(scanHtmlFormScopes('css`.a { }` + html`<row-btn></row-btn>`').tagUses, [{ tag: 'row-btn', scope: 'none', delivers: null, expr: null }]);
-  // A form written inside a plain string cannot open a scope for real markup.
-  const mixed = "const s = '<form>'; export default () => html`<button formaction=${del}>x</button>`;";
-  assert.deepEqual(scanHtmlFormScopes(mixed).submitters, [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }]);
-});
-
-test('an HTML comment is not markup', () => {
-  const src = 'html`<!-- <form> --><button formaction=${del}>x</button>`';
-  assert.deepEqual(scanHtmlFormScopes(src).submitters, [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }]);
-});
-
-test('classifyActionHole matches the tag and the attribute as a pair', () => {
-  assert.equal(classifyActionHole('<form action='), 'form');
-  assert.equal(classifyActionHole('<button formaction='), 'submitter');
-  assert.equal(classifyActionHole('<input formaction='), 'submitter');
-  assert.equal(classifyActionHole('<div action='), null, 'a div binds nothing');
-  assert.equal(classifyActionHole('<form formaction='), null, 'wrong attribute for the tag');
-  assert.equal(classifyActionHole('<button action='), null, 'wrong attribute for the tag');
-  assert.equal(classifyActionHole('<form action=x><span>'), null, 'the tag already closed');
-  assert.equal(classifyActionHole('plain text'), null);
-});
-
-test('matchClosingBrace walks past a template hole (#1307)', () => {
-  // A hole is a CODE context nested in a template, not a brace in the block
-  // being matched. Counting it toward the outer depth (the earlier behaviour)
-  // meant depth could never return to zero.
-  const s = '{ return html`<b>${x}</b>`; }';
-  assert.equal(matchClosingBrace(s, 1), s.length - 1);
-  const nested = '{ a(html`${ b(html`${c}`) }`); }';
-  assert.equal(matchClosingBrace(nested, 1), nested.length - 1);
-  // Still returns -1 when there really is no match.
-  assert.equal(matchClosingBrace('{ a(', 1), -1);
-});
-
-test('a class body holding a template hole is extractable from RAW source', () => {
-  // Every other caller passes a masked source in which holes are blanked, so
-  // this path was the one that exposed the brace bug.
-  const src = [
-    'class RowBtn extends WebComponent({}) {',
-    '  render() { return html`<button formaction=${del}>x</button>`; }',
-    '}',
-  ].join('\n');
-  const bodies = extractWebComponentClassBodies(src);
-  assert.equal(bodies.length, 1);
-  assert.match(bodies[0].body, /formaction=\$\{del\}/);
-  assert.deepEqual(scanHtmlFormScopes(bodies[0].body).submitters, [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }]);
-});
-
-test('an unbound form reports whether it would still DELIVER the identity', () => {
-  // This is the bit that decides broken from working, and it is not boundness.
-  // A submitter's identity rides its OWN name/value pair, so an unbound form
-  // that still sends a parseable POST body delivers it and the action runs.
-  const sub = (formTag) => scanHtmlFormScopes(
-    'html`' + formTag + '<button formaction=${del}>x</button></form>`',
-  ).submitters[0];
-
-  assert.equal(sub('<form method="post">').delivers, true, 'a POST body carries the pair');
-  assert.equal(sub('<form method="POST">').delivers, true, 'the keyword folds case');
-  assert.equal(sub('<form method="post" enctype="multipart/form-data">').delivers, true);
-  assert.equal(sub('<form>').delivers, false, 'no method is a GET');
-  assert.equal(sub('<form method="get">').delivers, false);
-  // `method` is an enumerated attribute matched against exact keywords with no
-  // whitespace stripping, so a padded value falls to the GET default.
-  assert.equal(sub('<form method=" post ">').delivers, false);
-  assert.equal(sub('<form method="post" enctype="text/plain">').delivers, false, 'the server cannot parse it');
-  // `enctype` is an enumerated attribute whose missing AND invalid value default
-  // are both application/x-www-form-urlencoded, so an unrecognised value falls
-  // back to a parseable body. Treating it as unparseable reported a working form
-  // as broken, which is the one thing this rule must never do.
-  assert.equal(sub('<form method="post" enctype="nonsense">').delivers, true, 'an invalid enctype falls back to urlencoded');
-  assert.equal(sub('<form method="post" enctype=" text/plain ">').delivers, true, 'padded, so invalid, so urlencoded');
-  assert.equal(sub('<form method="post" enctype="TEXT/PLAIN">').delivers, false, 'the keyword folds case');
-  // A hole anywhere else in the start tag makes the answer dynamic.
-  assert.equal(sub('<form method=${m}>').delivers, null);
-  assert.equal(sub('<form enctype=${e} method="post">').delivers, null);
-});
-
-test('opensForm reports whether the source opens any form at all', () => {
-  assert.equal(scanHtmlFormScopes('html`<button formaction=${d}>x</button>`').opensForm, false);
-  assert.equal(scanHtmlFormScopes('html`<form action=${s}></form>`').opensForm, true);
-  // A form written only inside a plain string is not a form.
-  assert.equal(scanHtmlFormScopes("const s = '<form>';").opensForm, false);
-});
-
-test('class-body offsets index the RAW source identically to the mask', () => {
-  // How the rule gets a body with its templates intact without asking the brace
-  // matcher to lex raw source: locate in the position-preserving mask, slice
-  // from `content`. A regex literal holding a brace is the case that broke when
-  // raw source was passed directly.
-  const src = [
-    'class RowBtn extends WebComponent({}) {',
-    '  static re = /[{]/;',
-    '  render() { return html`<button formaction=${del}>x</button>`; }',
-    '}',
-  ].join('\n');
-  const masked = redactStringsAndTemplates(src);
-  assert.equal(masked.length, src.length, 'the mask is position-preserving');
-  const bodies = extractWebComponentClassBodies(masked);
-  assert.equal(bodies.length, 1, 'the masked view blanks the regex body, so the braces balance');
-  const body = src.slice(bodies[0].bodyStart, bodies[0].bodyEnd);
-  assert.match(body, /formaction=\$\{del\}/, 'the raw slice keeps the template intact');
-  assert.deepEqual(scanHtmlFormScopes(body).submitters, [{ tag: 'button', scope: 'none', delivers: null, expr: 'del' }]);
-});
-
-test('a template in an ordinary START-TAG hole is handed off, not inherited', () => {
-  // A hole inside a start tag is an attribute or property VALUE whose placement
-  // this scan cannot speak for: SSR never renders it in place (a serializable
-  // value applies at hydration, a function-carrying one is dropped outright), so
-  // the receiving element decides where it lands in the browser. Scoring it by
-  // lexical nesting reported a shape the renderer treats as cannot-tell (and
-  // therefore binds) as a conclusive 'unbound'. `<webjs-suspense .fallback>` is
-  // the one exception and has its own test below.
-  const passed = 'html`<form method="post"><my-thing .tpl=${html`<button formaction=${del}>x</button>`}></my-thing></form>`';
-  // 'handed', NOT 'none'. Both mean "no enclosing form in this scan", but only
-  // 'none' is a cannot-tell the caller may attribute to this file's own
-  // component. A handed-off template belongs to whichever element received it,
-  // so collapsing the two resolves it through the wrong call sites.
-  assert.deepEqual(scanHtmlFormScopes(passed).submitters, [{ tag: 'button', scope: 'handed', delivers: null, expr: 'del' }]);
-  // A hyphenated tag inside a property hole is placed by the receiving element
-  // too, so it is not this file's call site either.
-  const tagPassed = 'html`<my-shell .rows=${html`<todo-row></todo-row>`}></my-shell>`';
-  assert.deepEqual(scanHtmlFormScopes(tagPassed).tagUses.filter((u) => u.tag === 'todo-row'),
-    [{ tag: 'todo-row', scope: 'handed', delivers: null, expr: null }]);
-  // A form the handed-off template opens ITSELF is still its own scope.
-  const ownForm = 'html`<my-thing .tpl=${html`<form action=${s}><button formaction=${del}>x</button></form>`}></my-thing>`';
-  assert.deepEqual(scanHtmlFormScopes(ownForm).submitters, [{ tag: 'button', scope: 'bound', delivers: null, expr: 'del' }]);
-  // A hole in CHILD position IS rendered inline by this scan, so it still
-  // inherits. This is the pair that keeps the fix from being a blanket opt-out.
-  const child = 'html`<form method="post">${html`<button formaction=${del}>x</button>`}</form>`';
-  assert.deepEqual(scanHtmlFormScopes(child).submitters, [{ tag: 'button', scope: 'unbound', delivers: true, expr: 'del' }]);
-});
-
-test('the unparseable-enctype constant tracks the renderer\'s own set', async () => {
-  // The scanner states its enctype rule as a DENYLIST of one because of the
-  // invalid-value default, while the renderer refuses the wider allowlist. This
-  // pins the relationship rather than asserting the two are equal, so a change
-  // to core's set surfaces here instead of drifting silently.
-  assert.ok(!PARSEABLE_ENCTYPES.has('text/plain'), 'the renderer cannot parse text/plain either');
-  for (const e of PARSEABLE_ENCTYPES) {
-    const src = `html\`<form method="post" enctype="${e}"><button formaction=\${d}>x</button></form>\``;
-    assert.equal(scanHtmlFormScopes(src).submitters[0].delivers, true, `${e} delivers`);
-  }
-  assert.deepEqual([...PARSEABLE_ENCTYPES].sort(),
-    ['application/x-www-form-urlencoded', 'multipart/form-data'],
-    'if core gains an enctype, revisit the scanner denylist');
-
-  // There are now THREE hardcoded copies of the denylist keyword: the scanner's
-  // `UNPARSEABLE_FORM_ENCTYPE`, the client guard in `router-client.js`, and this
-  // test. Pin the client one too, so the two halves of the feature cannot drift
-  // into disagreeing on the same input (which is what the allowlist did).
-  const clientSrc = await readFile(new URL('../../../core/src/router-client.js', import.meta.url), 'utf8');
-  assert.match(clientSrc, /enctype\.toLowerCase\(\) === 'text\/plain'/,
-    'the client guard uses the same one-keyword denylist, not the renderer allowlist');
-  assert.doesNotMatch(clientSrc, /PARSEABLE_ENCTYPES/,
-    'and does not reach for the renderer allowlist again');
-});
-
-test('a start-tag hole is handed off or inherited, and SSR agrees', async () => {
-  // A differential rather than an assertion about the scanner alone, because the
-  // whole 'handed' state is a claim about the renderer's behaviour. Most
-  // start-tag holes are values the receiving element places, but
-  // `<webjs-suspense .fallback=${…}>` is rendered INLINE with the enclosing form
-  // scope (#471: a custom-element property applies at hydration, too late for a
-  // placeholder that must be in the first flushed bytes), so the scanner reads
-  // it against that form rather than treating it as handed off.
-  const { html } = await import('../../../core/src/html.js');
-  const { renderToString } = await import('../../../core/src/render-server.js');
-  const { setFormActionResolver } = await import('../../../core/src/form-action.js');
-  setFormActionResolver(async () => 'abc1234567/publish');
-  const publish = async () => ({ success: true });
-
-  // Read against the enclosing form, not handed off. Nothing is asserted about
-  // a render-time refusal here: a bound submitter is SELF-SUFFICIENT (#1307), so
-  // the renderer stamps it with its own `formmethod="post"` and refuses only a
-  // same-element contradiction, never a cross-element one.
-  const fallbackSrc = 'html`<form><webjs-suspense .fallback=${html`<button formaction=${p}>x</button>`}></webjs-suspense></form>`';
-  assert.deepEqual(scanHtmlFormScopes(fallbackSrc).submitters,
-    [{ tag: 'button', scope: 'unbound', delivers: false, expr: 'p' }]);
-
-  // The negative half, and it has to assert the MECHANISM rather than "it did
-  // not throw". A template carrying a function cannot serialize, so SSR drops
-  // the whole property binding and emits nothing for it: the submitter is never
-  // rendered and never judged here at all. Asserting only that the render
-  // succeeded would stay green for ANY scanner verdict, which is precisely the
-  // non-discriminating test this differential exists to avoid.
-  const propSrc = 'html`<form><my-thing .tpl=${html`<button formaction=${p}>x</button>`}></my-thing></form>`';
-  assert.deepEqual(scanHtmlFormScopes(propSrc).submitters,
-    [{ tag: 'button', scope: 'handed', delivers: null, expr: 'p' }]);
-  const warns = [];
-  const quiet = console.warn;
-  console.warn = (...a) => warns.push(a.join(' '));
-  let out;
-  try {
-    out = await renderToString(
-      html`<form><my-thing .tpl=${html`<button formaction=${publish}>x</button>`}></my-thing></form>`,
-      { ssr: true },
-    );
-  } finally { console.warn = quiet; }
-  assert.doesNotMatch(out, /<button/, 'SSR emits nothing for the dropped property');
-  assert.ok(warns.some((w) => /unserializable value during SSR/.test(w)),
-    'and it says so, which is the proof SSR never judged this submitter');
-
-  // The tag half of the same carve-out. A hyphenated tag in a fallback IS a real
-  // call site (rendered inline, inside the enclosing form); one in an ordinary
-  // property hole is not.
-  const fbTag = 'html`<form><webjs-suspense .fallback=${html`<todo-row></todo-row>`}></webjs-suspense></form>`';
-  assert.deepEqual(scanHtmlFormScopes(fbTag).tagUses.filter((u) => u.tag === 'todo-row'),
-    [{ tag: 'todo-row', scope: 'unbound', delivers: false, expr: null }]);
-  const propTag = 'html`<form><my-shell .rows=${html`<todo-row></todo-row>`}></my-shell></form>`';
-  assert.deepEqual(scanHtmlFormScopes(propTag).tagUses.filter((u) => u.tag === 'todo-row'),
-    [{ tag: 'todo-row', scope: 'handed', delivers: null, expr: null }]);
-});
-
-test('the enctype divergence from the renderer is deliberate, on both sides', async () => {
-  // The scanner uses a one-keyword denylist and the renderer an allowlist, and
-  // it would be easy to "unify" them later. This pins WHY they differ, so that
-  // change breaks a test carrying its own reason.
-  //
-  // They ask different questions. This rule asks whether the identity ARRIVES,
-  // and an invalid enctype falls back to urlencoded, so it does. The renderer
-  // asks whether the form does what the author wrote, and an invalid value is
-  // the dangerous case: a typo'd `multipart/form-dat` falls back to urlencoded
-  // and silently drops every FILE from the submission.
-  const { html } = await import('../../../core/src/html.js');
-  const { renderToString } = await import('../../../core/src/render-server.js');
-  const { setFormActionResolver } = await import('../../../core/src/form-action.js');
-  setFormActionResolver(async () => 'abc1234567/save');
-  const save = async () => ({ success: true });
-
-  const sub = (formTag) => scanHtmlFormScopes(
-    'html`' + formTag + '<button formaction=${del}>x</button></form>`',
-  ).submitters[0];
-
-  // The scanner: an invalid enctype still delivers, so it is not a defect.
-  assert.equal(sub('<form method="post" enctype="multipart/form-dat">').delivers, true);
-  // The renderer: the same typo throws, because it would cost the author their
-  // file upload with no other signal.
-  await assert.rejects(
-    () => renderToString(html`<form action=${save} enctype=${'multipart/form-dat'}></form>`, { ssr: true }),
-    /cannot work|enctype/i,
-  );
-  // And both agree that text/plain is broken.
-  assert.equal(sub('<form method="post" enctype="text/plain">').delivers, false);
-  await assert.rejects(
-    () => renderToString(html`<form action=${save} enctype=${'text/plain'}></form>`, { ssr: true }),
-    /cannot work|enctype/i,
-  );
-});
